Eliminazione file deprecati
eliminazione file deprecati, divisione del file classi.php creando un file per ogni classe e inserendoli all'interno di una cartella, creata parte della classe orario
